requirements.txt
APScheduler==3.10.0
asgiref @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_cchykh5sgr/croots/recipe/asgiref_1661287542702/work
asttokens==2.1.0
attrs==22.2.0
backcall==0.2.0
cached-property==1.5.2
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==2.0.12
colorama==0.4.6
coverage==7.2.1
cryptography==39.0.1
cx-Oracle==8.3.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-filter==22.1
django-import-export==3.1.0
django-nested-admin==4.0.2
django-rest-auth==0.9.5
djangorestframework==3.13.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
idna==3.4
ipykernel==6.17.1
ipython==8.6.0
isodate==0.6.1
jedi==0.18.2
jupyter_client==7.4.7
jupyter_core==5.0.0
ldap3==2.9.1
lxml==4.9.2
MarkupPy==1.14
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.23.1
odfpy==1.4.1
openpyxl==3.1.1
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.5.4
prompt-toolkit==3.0.33
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.8
python-monkey-business==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.10.1
six==1.16.0
sqlparse @ file:///C:/b/abs_83jn6owfpc/croot/sqlparse_1668793499314/work
stack-data==0.6.1
tablib==3.3.0
tornado==6.2
traitlets==5.5.0
tzdata @ file:///tmp/build/80754af9/python-tzdata_1611693927741/work
tzlocal==4.2
urllib3==1.26.14
wcwidth==0.2.5
whitenoise==6.2.0
wincertstore==0.2
xlrd==2.0.1
XlsxWriter==3.0.8
xlwt==1.3.0
xmltodict==0.13.0
zeep==4.1.0
